refactor: split PrintSettings into subcomponents
