+++ src/utils/numberUtils.js
@@ -1,0 +1,1 @@
+export const clamp = (n, lo, hi) => Math.max(lo, Math.min(hi, n));
